requirements.txt
Django==4.2.0
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
gunicorn==21.2.0
